+++ pyccel/codegen/wrapper/fortran_to_c_wrapper.py
@@ -252,7 +252,7 @@ class FortranToCWrapper(Wrapper):
         func = BindCFunctionDef(name, func_arguments, body, FunctionDefResult(func_results), scope=func_scope, original_function = expr,
                 docstring = expr.docstring, result_pointer_map = expr.result_pointer_map)
 
-        self.scope.functions[name] = func
+        self.scope.insert_function(func, name)
 
         return func
 
@@ -273,8 +273,7 @@ class FortranToCWrapper(Wrapper):
         pyccel.ast.core.Interface
             The C-compatible interface.
         """
-        functions = [self.scope.functions[self._wrapper_names_dict[f.name]] for f in expr.functions]
-        functions = [f for f in functions if not isinstance(f, EmptyNode)]
+        functions = [self._wrap(f) for f in expr.functions if not isinstance(f, EmptyNode)]
         return Interface(expr.name, functions, expr.is_argument)
 
     def _extract_FunctionDefArgument(self, expr, func):
@@ -662,7 +661,7 @@ class FortranToCWrapper(Wrapper):
         BindCClassDef
             The wrapped class.
         """
-        name = expr.name
+        name = expr.scope.get_python_name(expr.name)
         func_name = self.scope.get_new_name(f'{name}_bind_c_alloc'.lower())
         func_scope = self.scope.new_child_scope(func_name)